refactor(services): update all imports to shared/ and remove shims
- Update all imports from old services paths to shared/
- Delete all re-export shim files
- Update test mock paths to issue.failure module
- Update services/__init__.py _SYMBOL_MODULES mapping

Breaking change: all imports must now use shared/ paths directly

Related to #2393 #2394 #2099
