+++ src/vibe3/services/flow/write_mixin.py
@@ -12,7 +12,7 @@ from vibe3.config import VibeConfig
 from vibe3.exceptions import UserError
 from vibe3.models import FlowState, FlowStatusResponse, MainBranchProtectedError
 from vibe3.services.flow.read_mixin import FlowReadMixin
-from vibe3.services.signature_service import SignatureService
+from vibe3.services.shared.signatures import SignatureService
 
 
 class FlowWriteMixin(FlowReadMixin):
