--- src/vibe3/services/label_service.py
@@ -1,9 +1,0 @@
-"""Re-export shim for label service.
-
-The actual implementation has moved to
-vibe3.services.shared.label_service.
-"""
-
-from vibe3.services.shared.label_service import LabelService
-
-__all__ = ["LabelService"]
