+++ src/vibe3/services/task/service.py
@@ -21,9 +21,9 @@ from vibe3.models import (
     PRResponse,
 )
 from vibe3.services.flow_service import FlowService
-from vibe3.services.label_service import LabelService
 from vibe3.services.pr.service import PRService
-from vibe3.services.signature_service import SignatureService
+from vibe3.services.shared.label_service import LabelService
+from vibe3.services.shared.signatures import SignatureService
 from vibe3.services.task.show import (
     TaskCommentSummary,
     TaskPRSummary,
