+++ tests/vibe3/execution/test_error_block_decoupling.py
@@ -100,7 +100,7 @@ class TestAgentNoopTriggersBlock:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_executor_noop_issue"
+                "vibe3.services.issue.failure.block_executor_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = {
@@ -138,7 +138,7 @@ class TestGitHubAPIFailureNoBlock:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_executor_noop_issue"
+                "vibe3.services.issue.failure.block_executor_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.side_effect = Exception("GitHub timeout")
