+++ tests/vibe3/execution/test_noop_gate_ref_check.py
@@ -39,7 +39,7 @@ class TestRefCheck:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_planner_noop_issue"
+                "vibe3.services.issue.failure.block_planner_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
@@ -68,7 +68,7 @@ class TestRefCheck:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_planner_noop_issue"
+                "vibe3.services.issue.failure.block_planner_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
@@ -96,7 +96,7 @@ class TestRefCheck:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_planner_noop_issue"
+                "vibe3.services.issue.failure.block_planner_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
@@ -123,7 +123,7 @@ class TestRefCheck:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_executor_noop_issue"
+                "vibe3.services.issue.failure.block_executor_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
@@ -150,7 +150,7 @@ class TestRefCheck:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_executor_noop_issue"
+                "vibe3.services.issue.failure.block_executor_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
@@ -176,7 +176,7 @@ class TestRefCheck:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_reviewer_noop_issue"
+                "vibe3.services.issue.failure.block_reviewer_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
@@ -212,7 +212,7 @@ class TestRefCheck:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_reviewer_noop_issue"
+                "vibe3.services.issue.failure.block_reviewer_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
@@ -248,7 +248,7 @@ class TestRefCheck:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_reviewer_noop_issue"
+                "vibe3.services.issue.failure.block_reviewer_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
@@ -274,7 +274,7 @@ class TestRefCheck:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_manager_noop_issue"
+                "vibe3.services.issue.failure.block_manager_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
