+++ tests/vibe3/roles/test_plan.py
@@ -32,7 +32,7 @@ class TestPlannerNoOpGate:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_planner_noop_issue"
+                "vibe3.services.issue.failure.block_planner_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = {
@@ -63,7 +63,7 @@ class TestPlannerNoOpGate:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_planner_noop_issue"
+                "vibe3.services.issue.failure.block_planner_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = {
