+++ tests/vibe3/roles/test_review.py
@@ -21,7 +21,7 @@ class TestReviewerNoOpGate:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_reviewer_noop_issue"
+                "vibe3.services.issue.failure.block_reviewer_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = {
@@ -65,7 +65,7 @@ class TestReviewerNoOpGate:
         with (
             patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
             patch(
-                "vibe3.services.role_policy_helpers.block_reviewer_noop_issue"
+                "vibe3.services.issue.failure.block_reviewer_noop_issue"
             ) as mock_block,
         ):
             mock_gh.return_value.view_issue.return_value = {
